Move Evo-only commands to `EvoWorklist` type
